Replace installer shell scripts with a Python program, which has better error handling and is easier to debug.  This also allows a single authorization prompt, since install/remove are back in one program.
Update checker can now be installed in home directory, so unprivileged users can now use the feature.

Calling code is untested in TLMMainWindowController.  Installer was tested from the command line, and is functional.

Need to ensure that the all users case is handled appropriately in the GUI, and possible conflicts are called out.